feat: identify IdPs using a certificate
AM-3274
