+++ gravitee-am-repository/gravitee-am-repository-api/src/main/java/io/gravitee/am/repository/management/api/IdentityProviderRepository.java
@@ -37,5 +37,4 @@ public interface IdentityProviderRepository extends CrudRepository<IdentityProvi
 
     Flowable<IdentityProvider> findAllByPasswordPolicy(ReferenceType referenceType, String referenceId, String passwordPolicy);
 
-    Flowable<IdentityProvider> findByCertificate(String domainId, String certId);
 }
