+++ gravitee-am-repository/gravitee-am-repository-jdbc/src/main/java/io/gravitee/am/repository/jdbc/management/api/JdbcIdentityProviderRepository.java
@@ -15,11 +15,7 @@
  */
 package io.gravitee.am.repository.jdbc.management.api;
 
-import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.node.ArrayNode;
-import com.fasterxml.jackson.databind.node.ObjectNode;
-import com.fasterxml.jackson.databind.node.TextNode;
 import io.gravitee.am.common.utils.RandomString;
 import io.gravitee.am.model.IdentityProvider;
 import io.gravitee.am.model.ReferenceType;
@@ -41,9 +37,6 @@ import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
-import java.util.Locale;
-import java.util.Map;
-import java.util.stream.Stream;
 
 import static java.util.Optional.ofNullable;
 import static reactor.adapter.rxjava.RxJava3Adapter.monoToSingle;
@@ -147,41 +140,6 @@ public class JdbcIdentityProviderRepository extends AbstractJdbcRepository imple
                 .map(this::toEntity);
     }
 
-    @Override
-    public Flowable<IdentityProvider> findByCertificate(String domainId, String certId) {
-        return findAll(ReferenceType.DOMAIN, domainId)
-                .filter(idp -> {
-                    // filter in-memory because 1) there shouldn't be that many idps in a domain
-                    // 2) we need custom lax logic for matching across all possible IDP config schemas
-                    var config = objectMapper.readTree(idp.getConfiguration());
-                    return hasEntryReferringToCert(config, certId);
-                });
-
-    }
-
-    private boolean hasEntryReferringToCert(JsonNode config, String certId) {
-        return config.properties()
-                .stream()
-                .anyMatch(entry -> refersToCert(entry, certId));
-    }
-
-    private boolean refersToCert(Map.Entry<String, JsonNode> entry, String certId) {
-        if (entry.getValue() instanceof ObjectNode nestedEntry) {
-            return hasEntryReferringToCert(nestedEntry, certId);
-        }
-        if (entry.getValue() instanceof ArrayNode arrayEntry) {
-            var elements = arrayEntry.elements();
-            return Stream.generate(elements::next)
-                    .anyMatch(elem -> refersToCert(Map.entry(entry.getKey(), elem), certId));
-        }
-        if (entry.getValue() instanceof TextNode textEntry) {
-            // various IDPs might use different keys for referring to a certificate, so we have to be lax
-            var isCertRelatedEntry = entry.getKey().toLowerCase(Locale.ROOT).contains("cert");
-            var matchesCertId = textEntry.textValue().equals(certId);
-            return isCertRelatedEntry && matchesCertId;
-        }
-        return false;
-    }
 
     @Override
     public Maybe<IdentityProvider> findById(String id) {
